pyproject.toml
[tool.poetry]
name = "scaffold"
version = "0.0.1"
description = "Boilerplate for building Discord bots in Python"
authors = ["Chase Franklin <gnullbyte@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
discord-py = "^1.3.3"
python-dotenv = "^0.13.0"
pynacl = "^1.4.0"
fuzzywuzzy = "^0.18.0"
cogs = "^0.4.4"
python-levenshtein = "^0.12.0"
dicey = "^2.0.6"

[tool.poetry.dev-dependencies]
pre-commit = "^2.5.1"
flake8 = "^3.8.3"
black = "^19.10b0"
twine = "^3.2.0"
wheel = "^0.34.2"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
